requirements.txt
pandas
python-dotenv
google-cloud-bigquery
db-dtypes
